@@ -17,13 +17,13 @@
     "build/*"
   ],
   "devDependencies": {
-    "eslint": "^4.5.0",
-    "eslint-config-synacor": "^1.1.0",
+    "eslint": "^5.12.1",
+    "eslint-config-synacor": "^3.0.3",
     "if-env": "^1.0.0",
     "preact-cli": "^2.0.0"
   },
   "dependencies": {
-    "preact": "^8.2.1",
-    "preact-compat": "^3.17.0"
+    "preact": "^8.4.2",
+    "preact-compat": "^3.18.4"
   }
 }
