@@ -71,199 +71,200 @@ module.exports =
 
 !function () {
   "use strict";
-  function e() {}function t(t, n) {
-    var o,
+  function e(e, t) {
+    var n,
+        o,
         r,
         i,
-        l,
-        a = E;for (l = arguments.length; l-- > 2;) {
-      W.push(arguments[l]);
-    }n && null != n.children && (W.length || W.push(n.children), delete n.children);while (W.length) {
-      if ((r = W.pop()) && void 0 !== r.pop) for (l = r.length; l--;) {
-        W.push(r[l]);
-      } else "boolean" == typeof r && (r = null), (i = "function" != typeof t) && (null == r ? r = "" : "number" == typeof r ? r += "" : "string" != typeof r && (i = !1)), i && o ? a[a.length - 1] += r : a === E ? a = [r] : a.push(r), o = i;
-    }var u = new e();return u.nodeName = t, u.children = a, u.attributes = null == n ? void 0 : n, u.key = null == n ? void 0 : n.key, void 0 !== S.vnode && S.vnode(u), u;
-  }function n(e, t) {
+        l = W;for (i = arguments.length; i-- > 2;) {
+      P.push(arguments[i]);
+    }t && null != t.children && (P.length || P.push(t.children), delete t.children);while (P.length) {
+      if ((o = P.pop()) && void 0 !== o.pop) for (i = o.length; i--;) {
+        P.push(o[i]);
+      } else "boolean" == typeof o && (o = null), (r = "function" != typeof e) && (null == o ? o = "" : "number" == typeof o ? o += "" : "string" != typeof o && (r = !1)), r && n ? l[l.length - 1] += o : l === W ? l = [o] : l.push(o), n = r;
+    }var a = new T();return a.nodeName = e, a.children = l, a.attributes = null == t ? void 0 : t, a.key = null == t ? void 0 : t.key, void 0 !== M.vnode && M.vnode(a), a;
+  }function t(e, t) {
     for (var n in t) {
       e[n] = t[n];
     }return e;
-  }function o(e, o) {
-    return t(e.nodeName, n(n({}, e.attributes), o), arguments.length > 2 ? [].slice.call(arguments, 2) : e.children);
+  }function n(e, t) {
+    null != e && ("function" == typeof e ? e(t) : e.current = t);
+  }function o(n, o) {
+    return e(n.nodeName, t(t({}, n.attributes), o), arguments.length > 2 ? [].slice.call(arguments, 2) : n.children);
   }function r(e) {
-    !e.__d && (e.__d = !0) && 1 == A.push(e) && (S.debounceRendering || P)(i);
+    !e.__d && (e.__d = !0) && 1 == V.push(e) && (M.debounceRendering || D)(i);
   }function i() {
-    var e,
-        t = A;A = [];while (e = t.pop()) {
-      e.__d && k(e);
+    var e;while (e = V.pop()) {
+      e.__d && x(e);
     }
   }function l(e, t, n) {
     return "string" == typeof t || "number" == typeof t ? void 0 !== e.splitText : "string" == typeof t.nodeName ? !e._componentConstructor && a(e, t.nodeName) : n || e._componentConstructor === t.nodeName;
   }function a(e, t) {
     return e.__n === t || e.nodeName.toLowerCase() === t.toLowerCase();
   }function u(e) {
-    var t = n({}, e.attributes);t.children = e.children;var o = e.nodeName.defaultProps;if (void 0 !== o) for (var r in o) {
-      void 0 === t[r] && (t[r] = o[r]);
-    }return t;
-  }function _(e, t) {
+    var n = t({}, e.attributes);n.children = e.children;var o = e.nodeName.defaultProps;if (void 0 !== o) for (var r in o) {
+      void 0 === n[r] && (n[r] = o[r]);
+    }return n;
+  }function c(e, t) {
     var n = t ? document.createElementNS("http://www.w3.org/2000/svg", e) : document.createElement(e);return n.__n = e, n;
   }function p(e) {
     var t = e.parentNode;t && t.removeChild(e);
-  }function c(e, t, n, o, r) {
-    if ("className" === t && (t = "class"), "key" === t) ;else if ("ref" === t) n && n(null), o && o(e);else if ("class" !== t || r) {
+  }function s(e, t, o, r, i) {
+    if ("className" === t && (t = "class"), "key" === t) ;else if ("ref" === t) n(o, null), n(r, e);else if ("class" !== t || i) {
       if ("style" === t) {
-        if (o && "string" != typeof o && "string" != typeof n || (e.style.cssText = o || ""), o && "object" == typeof o) {
-          if ("string" != typeof n) for (var i in n) {
-            i in o || (e.style[i] = "");
-          }for (var i in o) {
-            e.style[i] = "number" == typeof o[i] && !1 === V.test(i) ? o[i] + "px" : o[i];
+        if (r && "string" != typeof r && "string" != typeof o || (e.style.cssText = r || ""), r && "object" == typeof r) {
+          if ("string" != typeof o) for (var l in o) {
+            l in r || (e.style[l] = "");
+          }for (var l in r) {
+            e.style[l] = "number" == typeof r[l] && !1 === E.test(l) ? r[l] + "px" : r[l];
           }
         }
-      } else if ("dangerouslySetInnerHTML" === t) o && (e.innerHTML = o.__html || "");else if ("o" == t[0] && "n" == t[1]) {
-        var l = t !== (t = t.replace(/Capture$/, ""));t = t.toLowerCase().substring(2), o ? n || e.addEventListener(t, f, l) : e.removeEventListener(t, f, l), (e.__l || (e.__l = {}))[t] = o;
-      } else if ("list" !== t && "type" !== t && !r && t in e) s(e, t, null == o ? "" : o), null != o && !1 !== o || e.removeAttribute(t);else {
-        var a = r && t !== (t = t.replace(/^xlink\:?/, ""));null == o || !1 === o ? a ? e.removeAttributeNS("http://www.w3.org/1999/xlink", t.toLowerCase()) : e.removeAttribute(t) : "function" != typeof o && (a ? e.setAttributeNS("http://www.w3.org/1999/xlink", t.toLowerCase(), o) : e.setAttribute(t, o));
+      } else if ("dangerouslySetInnerHTML" === t) r && (e.innerHTML = r.__html || "");else if ("o" == t[0] && "n" == t[1]) {
+        var a = t !== (t = t.replace(/Capture$/, ""));t = t.toLowerCase().substring(2), r ? o || e.addEventListener(t, _, a) : e.removeEventListener(t, _, a), (e.__l || (e.__l = {}))[t] = r;
+      } else if ("list" !== t && "type" !== t && !i && t in e) {
+        try {
+          e[t] = null == r ? "" : r;
+        } catch (e) {}null != r && !1 !== r || "spellcheck" == t || e.removeAttribute(t);
+      } else {
+        var u = i && t !== (t = t.replace(/^xlink:?/, ""));null == r || !1 === r ? u ? e.removeAttributeNS("http://www.w3.org/1999/xlink", t.toLowerCase()) : e.removeAttribute(t) : "function" != typeof r && (u ? e.setAttributeNS("http://www.w3.org/1999/xlink", t.toLowerCase(), r) : e.setAttribute(t, r));
       }
-    } else e.className = o || "";
-  }function s(e, t, n) {
-    try {
-      e[t] = n;
-    } catch (e) {}
-  }function f(e) {
-    return this.__l[e.type](S.event && S.event(e) || e);
-  }function d() {
-    var e;while (e = D.pop()) {
-      S.afterMount && S.afterMount(e), e.componentDidMount && e.componentDidMount();
+    } else e.className = r || "";
+  }function _(e) {
+    return this.__l[e.type](M.event && M.event(e) || e);
+  }function f() {
+    var e;while (e = A.shift()) {
+      M.afterMount && M.afterMount(e), e.componentDidMount && e.componentDidMount();
     }
-  }function h(e, t, n, o, r, i) {
-    H++ || (R = null != r && void 0 !== r.ownerSVGElement, j = null != e && !("__preactattr_" in e));var l = m(e, t, n, o, i);return r && l.parentNode !== r && r.appendChild(l), --H || (j = !1, i || d()), l;
-  }function m(e, t, n, o, r) {
+  }function d(e, t, n, o, r, i) {
+    H++ || (R = null != r && void 0 !== r.ownerSVGElement, B = null != e && !("__preactattr_" in e));var l = h(e, t, n, o, i);return r && l.parentNode !== r && r.appendChild(l), --H || (B = !1, i || f()), l;
+  }function h(e, t, n, o, r) {
     var i = e,
-        l = R;if (null != t && "boolean" != typeof t || (t = ""), "string" == typeof t || "number" == typeof t) return e && void 0 !== e.splitText && e.parentNode && (!e._component || r) ? e.nodeValue != t && (e.nodeValue = t) : (i = document.createTextNode(t), e && (e.parentNode && e.parentNode.replaceChild(i, e), b(e, !0))), i.__preactattr_ = !0, i;var u = t.nodeName;if ("function" == typeof u) return U(e, t, n, o);if (R = "svg" === u || "foreignObject" !== u && R, u += "", (!e || !a(e, u)) && (i = _(u, R), e)) {
+        l = R;if (null != t && "boolean" != typeof t || (t = ""), "string" == typeof t || "number" == typeof t) return e && void 0 !== e.splitText && e.parentNode && (!e._component || r) ? e.nodeValue != t && (e.nodeValue = t) : (i = document.createTextNode(t), e && (e.parentNode && e.parentNode.replaceChild(i, e), v(e, !0))), i.__preactattr_ = !0, i;var u = t.nodeName;if ("function" == typeof u) return N(e, t, n, o);if (R = "svg" === u || "foreignObject" !== u && R, u += "", (!e || !a(e, u)) && (i = c(u, R), e)) {
       while (e.firstChild) {
         i.appendChild(e.firstChild);
-      }e.parentNode && e.parentNode.replaceChild(i, e), b(e, !0);
+      }e.parentNode && e.parentNode.replaceChild(i, e), v(e, !0);
     }var p = i.firstChild,
-        c = i.__preactattr_,
-        s = t.children;if (null == c) {
-      c = i.__preactattr_ = {};for (var f = i.attributes, d = f.length; d--;) {
-        c[f[d].name] = f[d].value;
+        s = i.__preactattr_,
+        _ = t.children;if (null == s) {
+      s = i.__preactattr_ = {};for (var f = i.attributes, d = f.length; d--;) {
+        s[f[d].name] = f[d].value;
       }
-    }return !j && s && 1 === s.length && "string" == typeof s[0] && null != p && void 0 !== p.splitText && null == p.nextSibling ? p.nodeValue != s[0] && (p.nodeValue = s[0]) : (s && s.length || null != p) && v(i, s, n, o, j || null != c.dangerouslySetInnerHTML), g(i, t.attributes, c), R = l, i;
-  }function v(e, t, n, o, r) {
+    }return !B && _ && 1 === _.length && "string" == typeof _[0] && null != p && void 0 !== p.splitText && null == p.nextSibling ? p.nodeValue != _[0] && (p.nodeValue = _[0]) : (_ && _.length || null != p) && m(i, _, n, o, B || null != s.dangerouslySetInnerHTML), y(i, t.attributes, s), R = l, i;
+  }function m(e, t, n, o, r) {
     var i,
         a,
         u,
-        _,
         c,
-        s = e.childNodes,
+        s,
+        _ = e.childNodes,
         f = [],
         d = {},
-        h = 0,
-        v = 0,
-        y = s.length,
+        m = 0,
+        b = 0,
+        y = _.length,
         g = 0,
         w = t ? t.length : 0;if (0 !== y) for (var C = 0; C < y; C++) {
-      var x = s[C],
+      var x = _[C],
           N = x.__preactattr_,
-          k = w && N ? x._component ? x._component.__k : N.key : null;null != k ? (h++, d[k] = x) : (N || (void 0 !== x.splitText ? !r || x.nodeValue.trim() : r)) && (f[g++] = x);
+          k = w && N ? x._component ? x._component.__k : N.key : null;null != k ? (m++, d[k] = x) : (N || (void 0 !== x.splitText ? !r || x.nodeValue.trim() : r)) && (f[g++] = x);
     }if (0 !== w) for (var C = 0; C < w; C++) {
-      _ = t[C], c = null;var k = _.key;if (null != k) h && void 0 !== d[k] && (c = d[k], d[k] = void 0, h--);else if (!c && v < g) for (i = v; i < g; i++) {
-        if (void 0 !== f[i] && l(a = f[i], _, r)) {
-          c = a, f[i] = void 0, i === g - 1 && g--, i === v && v++;break;
+      c = t[C], s = null;var k = c.key;if (null != k) m && void 0 !== d[k] && (s = d[k], d[k] = void 0, m--);else if (b < g) for (i = b; i < g; i++) {
+        if (void 0 !== f[i] && l(a = f[i], c, r)) {
+          s = a, f[i] = void 0, i === g - 1 && g--, i === b && b++;break;
         }
-      }c = m(c, _, n, o), u = s[C], c && c !== e && c !== u && (null == u ? e.appendChild(c) : c === u.nextSibling ? p(u) : e.insertBefore(c, u));
-    }if (h) for (var C in d) {
-      void 0 !== d[C] && b(d[C], !1);
-    }while (v <= g) {
-      void 0 !== (c = f[g--]) && b(c, !1);
+      }s = h(s, c, n, o), u = _[C], s && s !== e && s !== u && (null == u ? e.appendChild(s) : s === u.nextSibling ? p(u) : e.insertBefore(s, u));
+    }if (m) for (var C in d) {
+      void 0 !== d[C] && v(d[C], !1);
+    }while (b <= g) {
+      void 0 !== (s = f[g--]) && v(s, !1);
     }
-  }function b(e, t) {
-    var n = e._component;n ? L(n) : (null != e.__preactattr_ && e.__preactattr_.ref && e.__preactattr_.ref(null), !1 !== t && null != e.__preactattr_ || p(e), y(e));
-  }function y(e) {
+  }function v(e, t) {
+    var o = e._component;o ? k(o) : (null != e.__preactattr_ && n(e.__preactattr_.ref, null), !1 !== t && null != e.__preactattr_ || p(e), b(e));
+  }function b(e) {
     e = e.lastChild;while (e) {
-      var t = e.previousSibling;b(e, !0), e = t;
+      var t = e.previousSibling;v(e, !0), e = t;
+    }
+  }function y(e, t, n) {
+    var o;for (o in n) {
+      t && null != t[o] || null == n[o] || s(e, o, n[o], n[o] = void 0, R);
+    }for (o in t) {
+      "children" === o || "innerHTML" === o || o in n && t[o] === ("value" === o || "checked" === o ? e[o] : n[o]) || s(e, o, n[o], n[o] = t[o], R);
     }
   }function g(e, t, n) {
-    var o;for (o in n) {
-      t && null != t[o] || null == n[o] || c(e, o, n[o], n[o] = void 0, R);
-    }for (o in t) {
-      "children" === o || "innerHTML" === o || o in n && t[o] === ("value" === o || "checked" === o ? e[o] : n[o]) || c(e, o, n[o], n[o] = t[o], R);
-    }
-  }function w(e) {
-    var t = e.constructor.name;(I[t] || (I[t] = [])).push(e);
-  }function C(e, t, n) {
     var o,
-        r = I[e.name];if (e.prototype && e.prototype.render ? (o = new e(t, n), T.call(o, t, n)) : (o = new T(t, n), o.constructor = e, o.render = x), r) for (var i = r.length; i--;) {
-      if (r[i].constructor === e) {
-        o.__b = r[i].__b, r.splice(i, 1);break;
-      }
+        r = F.length;e.prototype && e.prototype.render ? (o = new e(t, n), U.call(o, t, n)) : (o = new U(t, n), o.constructor = e, o.render = w);while (r--) {
+      if (F[r].constructor === e) return o.__b = F[r].__b, F.splice(r, 1), o;
     }return o;
-  }function x(e, t, n) {
+  }function w(e, t, n) {
     return this.constructor(e, n);
-  }function N(e, t, n, o, i) {
-    e.__x || (e.__x = !0, (e.__r = t.ref) && delete t.ref, (e.__k = t.key) && delete t.key, !e.base || i ? e.componentWillMount && e.componentWillMount() : e.componentWillReceiveProps && e.componentWillReceiveProps(t, o), o && o !== e.context && (e.__c || (e.__c = e.context), e.context = o), e.__p || (e.__p = e.props), e.props = t, e.__x = !1, 0 !== n && (1 !== n && !1 === S.syncComponentUpdates && e.base ? r(e) : k(e, 1, i)), e.__r && e.__r(e));
-  }function k(e, t, o, r) {
+  }function C(e, t, o, i, l) {
+    e.__x || (e.__x = !0, e.__r = t.ref, e.__k = t.key, delete t.ref, delete t.key, void 0 === e.constructor.getDerivedStateFromProps && (!e.base || l ? e.componentWillMount && e.componentWillMount() : e.componentWillReceiveProps && e.componentWillReceiveProps(t, i)), i && i !== e.context && (e.__c || (e.__c = e.context), e.context = i), e.__p || (e.__p = e.props), e.props = t, e.__x = !1, 0 !== o && (1 !== o && !1 === M.syncComponentUpdates && e.base ? r(e) : x(e, 1, l)), n(e.__r, e));
+  }function x(e, n, o, r) {
     if (!e.__x) {
       var i,
           l,
           a,
-          _ = e.props,
+          c = e.props,
           p = e.state,
-          c = e.context,
-          s = e.__p || _,
-          f = e.__s || p,
-          m = e.__c || c,
-          v = e.base,
+          s = e.context,
+          _ = e.__p || c,
+          h = e.__s || p,
+          m = e.__c || s,
+          b = e.base,
           y = e.__b,
-          g = v || y,
-          w = e._component,
-          x = !1;if (v && (e.props = s, e.state = f, e.context = m, 2 !== t && e.shouldComponentUpdate && !1 === e.shouldComponentUpdate(_, p, c) ? x = !0 : e.componentWillUpdate && e.componentWillUpdate(_, p, c), e.props = _, e.state = p, e.context = c), e.__p = e.__s = e.__c = e.__b = null, e.__d = !1, !x) {
-        i = e.render(_, p, c), e.getChildContext && (c = n(n({}, c), e.getChildContext()));var U,
+          w = b || y,
+          N = e._component,
+          U = !1,
+          S = m;if (e.constructor.getDerivedStateFromProps && (p = t(t({}, p), e.constructor.getDerivedStateFromProps(c, p)), e.state = p), b && (e.props = _, e.state = h, e.context = m, 2 !== n && e.shouldComponentUpdate && !1 === e.shouldComponentUpdate(c, p, s) ? U = !0 : e.componentWillUpdate && e.componentWillUpdate(c, p, s), e.props = c, e.state = p, e.context = s), e.__p = e.__s = e.__c = e.__b = null, e.__d = !1, !U) {
+        i = e.render(c, p, s), e.getChildContext && (s = t(t({}, s), e.getChildContext())), b && e.getSnapshotBeforeUpdate && (S = e.getSnapshotBeforeUpdate(_, h));var L,
             T,
-            M = i && i.nodeName;if ("function" == typeof M) {
-          var W = u(i);l = w, l && l.constructor === M && W.key == l.__k ? N(l, W, 1, c, !1) : (U = l, e._component = l = C(M, W, c), l.__b = l.__b || y, l.__u = e, N(l, W, 0, c, !1), k(l, 1, o, !0)), T = l.base;
-        } else a = g, U = w, U && (a = e._component = null), (g || 1 === t) && (a && (a._component = null), T = h(a, i, c, o || !v, g && g.parentNode, !0));if (g && T !== g && l !== w) {
-          var E = g.parentNode;E && T !== E && (E.replaceChild(T, g), U || (g._component = null, b(g, !1)));
-        }if (U && L(U), e.base = T, T && !r) {
-          var P = e,
+            P = i && i.nodeName;if ("function" == typeof P) {
+          var W = u(i);l = N, l && l.constructor === P && W.key == l.__k ? C(l, W, 1, s, !1) : (L = l, e._component = l = g(P, W, s), l.__b = l.__b || y, l.__u = e, C(l, W, 0, s, !1), x(l, 1, o, !0)), T = l.base;
+        } else a = w, L = N, L && (a = e._component = null), (w || 1 === n) && (a && (a._component = null), T = d(a, i, s, o || !b, w && w.parentNode, !0));if (w && T !== w && l !== N) {
+          var D = w.parentNode;D && T !== D && (D.replaceChild(T, w), L || (w._component = null, v(w, !1)));
+        }if (L && k(L), e.base = T, T && !r) {
+          var E = e,
               V = e;while (V = V.__u) {
-            (P = V).base = T;
-          }T._component = P, T._componentConstructor = P.constructor;
+            (E = V).base = T;
+          }T._component = E, T._componentConstructor = E.constructor;
         }
-      }if (!v || o ? D.unshift(e) : x || (e.componentDidUpdate && e.componentDidUpdate(s, f, m), S.afterUpdate && S.afterUpdate(e)), null != e.__h) while (e.__h.length) {
+      }!b || o ? A.push(e) : U || (e.componentDidUpdate && e.componentDidUpdate(_, h, S), M.afterUpdate && M.afterUpdate(e));while (e.__h.length) {
         e.__h.pop().call(e);
-      }H || r || d();
+      }H || r || f();
     }
-  }function U(e, t, n, o) {
+  }function N(e, t, n, o) {
     var r = e && e._component,
         i = r,
         l = e,
         a = r && e._componentConstructor === t.nodeName,
-        _ = a,
-        p = u(t);while (r && !_ && (r = r.__u)) {
-      _ = r.constructor === t.nodeName;
-    }return r && _ && (!o || r._component) ? (N(r, p, 3, n, o), e = r.base) : (i && !a && (L(i), e = l = null), r = C(t.nodeName, p, n), e && !r.__b && (r.__b = e, l = null), N(r, p, 1, n, o), e = r.base, l && e !== l && (l._component = null, b(l, !1))), e;
-  }function L(e) {
-    S.beforeUnmount && S.beforeUnmount(e);var t = e.base;e.__x = !0, e.componentWillUnmount && e.componentWillUnmount(), e.base = null;var n = e._component;n ? L(n) : t && (t.__preactattr_ && t.__preactattr_.ref && t.__preactattr_.ref(null), e.__b = t, p(t), w(e), y(t)), e.__r && e.__r(null);
-  }function T(e, t) {
-    this.__d = !0, this.context = t, this.props = e, this.state = this.state || {};
-  }function M(e, t, n) {
-    return h(n, e, {}, !1, t, !1);
-  }var S = {},
+        c = a,
+        p = u(t);while (r && !c && (r = r.__u)) {
+      c = r.constructor === t.nodeName;
+    }return r && c && (!o || r._component) ? (C(r, p, 3, n, o), e = r.base) : (i && !a && (k(i), e = l = null), r = g(t.nodeName, p, n), e && !r.__b && (r.__b = e, l = null), C(r, p, 1, n, o), e = r.base, l && e !== l && (l._component = null, v(l, !1))), e;
+  }function k(e) {
+    M.beforeUnmount && M.beforeUnmount(e);var t = e.base;e.__x = !0, e.componentWillUnmount && e.componentWillUnmount(), e.base = null;var o = e._component;o ? k(o) : t && (null != t.__preactattr_ && n(t.__preactattr_.ref, null), e.__b = t, p(t), F.push(e), b(t)), n(e.__r, null);
+  }function U(e, t) {
+    this.__d = !0, this.context = t, this.props = e, this.state = this.state || {}, this.__h = [];
+  }function S(e, t, n) {
+    return d(n, e, {}, !1, t, !1);
+  }function L() {
+    return {};
+  }var T = function T() {},
+      M = {},
+      P = [],
       W = [],
-      E = [],
-      P = "function" == typeof Promise ? Promise.resolve().then.bind(Promise.resolve()) : setTimeout,
-      V = /acit|ex(?:s|g|n|p|$)|rph|ows|mnc|ntw|ine[ch]|zoo|^ord/i,
+      D = "function" == typeof Promise ? Promise.resolve().then.bind(Promise.resolve()) : setTimeout,
+      E = /acit|ex(?:s|g|n|p|$)|rph|ows|mnc|ntw|ine[ch]|zoo|^ord/i,
+      V = [],
       A = [],
-      D = [],
       H = 0,
       R = !1,
-      j = !1,
-      I = {};n(T.prototype, { setState: function setState(e, t) {
-      var o = this.state;this.__s || (this.__s = n({}, o)), n(o, "function" == typeof e ? e(o, this.props) : e), t && (this.__h = this.__h || []).push(t), r(this);
+      B = !1,
+      F = [];t(U.prototype, { setState: function setState(e, n) {
+      this.__s || (this.__s = this.state), this.state = t(t({}, this.state), "function" == typeof e ? e(this.state, this.props) : e), n && this.__h.push(n), r(this);
     }, forceUpdate: function forceUpdate(e) {
-      e && (this.__h = this.__h || []).push(e), k(this, 2);
-    }, render: function render() {} });var $ = { h: t, createElement: t, cloneElement: o, Component: T, render: M, rerender: i, options: S }; true ? module.exports = $ : self.preact = $;
+      e && this.__h.push(e), x(this, 2);
+    }, render: function render() {} });var j = { h: e, createElement: e, cloneElement: o, createRef: L, Component: U, render: S, rerender: i, options: M }; true ? module.exports = j : self.preact = j;
 }();
 //# sourceMappingURL=preact.min.js.map
 
